requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Authlib==1.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.24.66
botocore==1.27.66
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
dateparser==1.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-4-jet==1.0.9
django-cors-headers==3.13.0
django-logging-middleware==0.1.2
django-ses==3.1.2
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
greenlet==1.1.3
idna==3.3
jet-bridge-base==1.2.7
jet-django==1.4.0
jmespath==1.0.1
kombu==5.2.4
loguru==0.6.0
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==2.0.9
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-core==4.3.0
SQLAlchemy==1.4.41
sqlparse==0.2.4
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
mongoengine==0.24.2
pymongo==3.12.1
djongo==1.3.6
django-celery-beat==2.4.0